Generalized the filter
